Using hooks for more stable restriction!
Now hijacking the windows mouse DLL Call to force restrict the mouse without having to "fight" windowses handling of it.
This came with a complete restructure of the code, we are no longer using a 'main loop' where all the logic is handled,
Instead we are using hotkeys. thanks to @nnnik#6686  on discord for help with that.

Overall, better performance, better user experience.   Enjoy!